+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   EUR and GDP currency symbols.
+-   Exchange rate fetching for all supported currencies.
 
 ## [0.17.0] - 2026-01-13